python-spring-equivalent/requirements.txt
# FastAPI and web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.0

# JWT and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# OpenTelemetry and monitoring
opentelemetry-api==1.28.0
opentelemetry-sdk==1.28.0
opentelemetry-instrumentation-fastapi==0.49b0
opentelemetry-instrumentation-requests==0.49b0
opentelemetry-instrumentation-httpx==0.49b0
opentelemetry-exporter-otlp==1.28.0
opentelemetry-instrumentation-logging==0.49b0

# HTTP client
httpx==0.28.1
requests==2.32.3

# Database and caching
redis==5.2.1
motor==3.6.0  # Async MongoDB driver

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1

# Code quality
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.13.0
bandit==1.7.10
safety==3.2.0

# Documentation
sphinx==8.1.3
sphinx-rtd-theme==3.0.2

# Health checks
psutil==6.1.0

# Logging
structlog==24.4.0
